expression to module
